+++ src/gh_profiler/utils/analysis_utils.py
@@ -9,6 +9,11 @@ from . import flags
 
 def process_data():
     """Process all data."""
+    if pdata.username == "ghost":
+        # This is GitHub's deleted user, and we don't need to do anything.
+        pdata.flag_overall_profile = flags.red_flag
+        return
+        
     _process_account_age()
     _process_profile_info()
     _process_pr_activity()