+++ src/gh_profiler/utils/flags.py
@@ -3,3 +3,5 @@
 red_flag = "\U0001f534"
 yellow_flag = "\U0001f7e1"
 green_flag = "\U0001f7e2"
+
+merged_flag = "\U0001f7e3"