+++ src/gh_profiler/utils/repo_data.py
@@ -1,6 +1,7 @@
 """One place to store all data about the targeted repo."""
 
 from dataclasses import dataclass
+from datetime import datetime as dt, timezone
 
 
 @dataclass(slots=True)
@@ -20,9 +21,20 @@ class PRData:
     author: str = ""
     title: str = ""
     url: str = ""
+    
+    merged: bool | None = None
+    closed_at: dt | None = None
+
 
     # Processed info
+    author_summary: str = ""
     summary: str = ""
+
+    # User flags
+    # When pdata is not a singleton, we shouldn't need to store these here.
+    profile_flag: str = ""
+    pr_flag: str = ""
+    issue_flag: str = ""
 
 
 repo_data = RepoData()