+++ tests/e2e_tests/test_profiler.py
@@ -137,7 +137,7 @@ def test_bulk_open_prs():
     cmd = f"uv run gh-profiler {url} -n 3"
     output = run_with_timeout(cmd)
 
-    assert output.count("https://github.com/django/django/pull/") == 3
+    assert output.count("https://github.com/django/django/pull/") == 6
     assert output.count("GitHub user: ") == 3
 
     # Check that there are 3 PR numbers and titles in output.
@@ -145,3 +145,75 @@ def test_bulk_open_prs():
     matches = re.findall(re_pr_title, output)
     assert len(matches) == 3
 
+
+def test_bulk_open_prs_table_only():
+    """Test a table-only run against a repo URL for bulk processing open PRs."""
+    # Django is likely to have many open PRs.
+    url = "https://github.com/django/django"
+    cmd = f"uv run gh-profiler {url} -n 3 --table-only"
+    output = run_with_timeout(cmd)
+
+    assert output.count("https://github.com/django/django/pull/") == 3
+    assert "GitHub user: " not in output
+    assert "Merged." not in output
+    assert "Closed." not in output
+
+    # Make sure Merged? column is not in output for open PRs.
+    assert "Merged?" not in output
+
+    # Check that there are no PR numbers and titles in output.
+    re_pr_title = r"(PR \d+: ).*"
+    matches = re.findall(re_pr_title, output)
+    assert len(matches) == 0
+
+
+def test_bulk_closed_prs():
+    """Test a run against a repo URL for bulk processing closed PRs."""
+    # Django is likely to have many closed PRs.
+    url = "https://github.com/django/django"
+    cmd = f"uv run gh-profiler {url} --back -n 3"
+    output = run_with_timeout(cmd)
+
+    assert output.count("https://github.com/django/django/pull/") == 6
+    assert output.count("GitHub user: ") + output.count("`ghost`") == 3
+    assert output.count("Merged.") + output.count("Closed.") == 3
+
+    # Check that there are 3 PR numbers and titles in output.
+    re_pr_title = r"(PR \d+: ).*"
+    matches = re.findall(re_pr_title, output)
+    assert len(matches) == 3
+
+
+def test_bulk_closed_prs_table_only():
+    """Test a table-only run against a repo URL for bulk processing closed PRs."""
+    # Django is likely to have many closed PRs.
+    url = "https://github.com/django/django"
+    cmd = f"uv run gh-profiler {url} --back -n 3 --table-only"
+    output = run_with_timeout(cmd)
+
+    assert output.count("https://github.com/django/django/pull/") == 3
+    assert "GitHub user: " not in output
+    assert "Merged." not in output
+    assert "Closed." not in output
+
+    # Check that there are no PR numbers and titles in output.
+    re_pr_title = r"(PR \d+: ).*"
+    matches = re.findall(re_pr_title, output)
+    assert len(matches) == 0
+
+
+@pytest.mark.parametrize("mode", ["", "--concise"])
+def test_ghost_profile(mode):
+    """Test a run against the `ghost` user.
+
+    GitHub has a special user account named `ghost` that stands in for
+    deleted users in completed actions, such as closed PRs.
+
+    This should return a single red flag, with an explanation of this role.
+
+    See: https://github.com/ghost
+    """
+    cmd = f"uv run gh-profiler ghost {mode}"
+    output = run_with_timeout(cmd)
+
+    assert output.strip() == f"🔴 The `ghost` account is GitHub's reference to a deleted user."